DOES NOT BUILD!!
Moved filter and type classes to the new Sidekick.Business project.
Renamed LanguageService to LanguageProvider.
Adjusted Languages namespace.
