+++ Sidekick.Business/Filters/FilterOption.cs
@@ -1,0 +1,7 @@
+namespace Sidekick.Business.Filters
+{
+    public class FilterOption
+    {
+        public string Option { get; set; }
+    }
+}
