+++ Sidekick.Business/Filters/FilterValue.cs
@@ -1,0 +1,8 @@
+namespace Sidekick.Business.Filters
+{
+    public class FilterValue
+    {
+        public int? Min { get; set; }
+        public int? Max { get; set; }
+    }
+}
