+++ Sidekick.Business/Filters/MiscFilter.cs
@@ -1,0 +1,8 @@
+namespace Sidekick.Business.Filters
+{
+    public class MiscFilter
+    {
+        public bool Disabled { get; set; }
+        public MiscFilters Filters { get; set; } = new MiscFilters();
+    }
+}
