+++ Sidekick.Business/Filters/RequirementFilter.cs
@@ -1,0 +1,8 @@
+namespace Sidekick.Business.Filters
+{
+    public class RequierementFilter
+    {
+        public bool Disabled { get; set; }
+        public RequierementFilters Filters { get; set; } = new RequierementFilters();
+    }
+}
