+++ Sidekick.Business/Filters/SocketFilters.cs
@@ -1,0 +1,8 @@
+namespace Sidekick.Business.Filters
+{
+    public class SocketFilters
+    {
+        public SocketFilterOption Sockets { get; set; }
+        public SocketFilterOption Links { get; set; }
+    }
+}
