+++ Sidekick.Business/Filters/StatFilter.cs
@@ -1,0 +1,9 @@
+namespace Sidekick.Business.Filters
+{
+    public class StatFilter
+    {
+        public string Id { get; set; }
+        public FilterValue Value { get; set; }
+        public bool Disabled { get; set; }
+    }
+}
