+++ Sidekick.Business/Filters/Status.cs
@@ -1,0 +1,9 @@
+using Sidekick.Business.Types;
+
+namespace Sidekick.Business.Filters
+{
+    public class Status
+    {
+        public StatusType Option { get; set; }
+    }
+}
