+++ Sidekick.Business/Filters/TypeFilters.cs
@@ -1,0 +1,8 @@
+namespace Sidekick.Business.Filters
+{
+    public class TypeFilters
+    {
+        public FilterOption Category { get; set; }
+        public FilterOption Rarity { get; set; }
+    }
+}
