+++ Sidekick.Business/Filters/WeaponFilter.cs
@@ -1,0 +1,8 @@
+namespace Sidekick.Business.Filters
+{
+    public class WeaponFilter
+    {
+        public bool Disabled { get; set; }
+        public WeaponFilters Filters { get; set; } = new WeaponFilters();
+    }
+}
