+++ Sidekick.Business/Filters/WeaponFilters.cs
@@ -1,0 +1,11 @@
+namespace Sidekick.Business.Filters
+{
+    public class WeaponFilters
+    {
+        public FilterValue Damage { get; set; }
+        public FilterValue Crit { get; set; }
+        public FilterValue APS { get; set; }
+        public FilterValue EDPS { get; set; }
+        public FilterValue PDPS { get; set; }
+    }
+}
