+++ Sidekick.Business/Types/InfluenceType.cs
@@ -1,0 +1,13 @@
+namespace Sidekick.Business.Types
+{
+    public enum InfluenceType
+    {
+        None,
+        Shaper,
+        Elder,
+        Crusader,
+        Hunter,
+        Redeemer,
+        Warlord,
+    };
+}
