+++ Sidekick.Business/Types/SortType.cs
@@ -1,0 +1,8 @@
+namespace Sidekick.Business.Types
+{
+    public enum SortType
+    {
+        Asc,
+        Desc
+    }
+}
