+++ Sidekick.Business/Types/StatType.cs
@@ -1,0 +1,9 @@
+namespace Sidekick.Business.Types
+{
+    public enum StatType
+    {
+        And,
+        Or,
+        Count
+    }
+}
