+++ Sidekick.Business/Types/StatusType.cs
@@ -1,0 +1,8 @@
+namespace Sidekick.Business.Types
+{
+    public enum StatusType
+    {
+        Online,
+        Any
+    }
+}
